+++ bot/verify.js
@@ -32,32 +32,150 @@ async function sendToChannels(channels, body) {
       discordFetch(`/channels/${ch.id}/messages`, {
         method: 'POST',
         body: JSON.stringify(body),
-      })
-    )
+      }),
+    ),
   );
   return results.filter((r) => r.status === 'fulfilled').length;
 }
 
-async function sendLaunchMessage(guildId, eventId, eventName, teams) {
+async function sendLaunchMessage(guildId, eventId, eventName, teams, startDate) {
   const channels = await getEventChannels(guildId, eventId);
-  if (channels.length === 0) return { success: false, error: 'No channels found with event ID in topic' };
+  if (channels.length === 0)
+    return { success: false, error: 'No channels found with event ID in topic' };
+
+  const eventUrl = `${FRONTEND_URL}/gielinor-rush/${eventId}`;
+  const teamList = teams.map((t) => `\`${t.teamName}\``).join('  ·  ') || 'TBD';
+
+  const startTs = startDate ? Math.floor(new Date(startDate).getTime() / 1000) : null;
+  const hasStarted = !startTs || Date.now() / 1000 >= startTs;
+
+  const embed = {
+    title: `🏆  ${eventName}  🏆`,
+    color: hasStarted ? 0xf0c040 : 0x5865f2,
+    description: hasStarted
+      ? [
+          `# 🗺️ The Gielinor Rush is live!`,
+          ``,
+          `> Grab your gear, rally your crew, and race across Gielinor.`,
+          `> The map is open and glory awaits. May the best team win!!`,
+          ``,
+          `## **[🔗 Open Event Overview →](${eventUrl})**`,
+        ].join('\n')
+      : [
+          `# 📅 The Gielinor Rush is set up and ready to go!`,
+          ``,
+          `> **Submissions are not open yet.** You won't be able to submit anything until the event officially starts.`,
+          ``,
+          `## ⏰  Event starts <t:${startTs}:F> — <t:${startTs}:R>`,
+          ``,
+          `## **[🔗 Explore the Event Overview →](${eventUrl})**`,
+        ].join('\n'),
+    fields: [
+      { name: '🆔  Event ID', value: `\`\`\`${eventId}\`\`\``, inline: true },
+      { name: '👥  Competing Teams', value: teamList, inline: true },
+      { name: '​', value: '────────────────────────────────────────' },
+      {
+        name: hasStarted ? '⚡  Getting Started' : '🗺️  What you can do now',
+        value: hasStarted
+          ? [
+              `> **1)** [View the map & leaderboard](${eventUrl})`,
+              `> **2)** Head to your team page (linked from the overview)`,
+              `> **3)** Link your Discord on your [**OSRS Bingo Hub**](${FRONTEND_URL}) profile to unlock inn visits & purchases`,
+              `> **4)** Have teammates submit a test screenshot for the **start node** so everyone knows the flow`,
+            ].join('\n')
+          : [
+              `> **1)** [Browse the event overview & map](${eventUrl})`,
+              `> **2)** Find your team page and get familiar with the layout`,
+              `> **3)** Link your Discord on your [**OSRS Bingo Hub**](${FRONTEND_URL}) profile now so you're ready to go`,
+              `> **4)** Read the rules -- submissions open the moment the clock hits <t:${startTs}:t>`,
+            ].join('\n'),
+      },
+      { name: '​', value: '────────────────────────────────────────' },
+      {
+        name: '📜  READ THE RULES! Yes, really.',
+        value: [
+          `> ⚠️ Go to the [event page](${eventUrl}) and read **every single word**.`,
+          `> ...`,
+          `> **Please. For the sake of the event admins. :)**`,
+        ].join('\n'),
+      },
+    ],
+    footer: {
+      text: hasStarted
+        ? '🏅 Good luck, and happy scaping!'
+        : '⏳ Sit tight — the race begins soon!',
+    },
+    timestamp: new Date().toISOString(),
+  };
+
+  const sent = await sendToChannels(channels, { embeds: [embed] });
+  return { success: sent > 0, channelsSent: sent };
+}
+
+async function sendCompleteMessage(guildId, eventId, eventName, teams) {
+  const channels = await getEventChannels(guildId, eventId);
+  if (channels.length === 0)
+    return { success: false, error: 'No channels found with event ID in topic' };
+
+  const eventUrl = `${FRONTEND_URL}/gielinor-rush/${eventId}`;
+  const medals = ['🥇', '🥈', '🥉'];
+  const formatGp = (gpStr) => BigInt(gpStr || '0').toLocaleString();
+
+  const sorted = [...(teams || [])].sort((a, b) =>
+    BigInt(b.currentPot || '0') > BigInt(a.currentPot || '0') ? 1 : -1,
+  );
+
+  const standingsLines = sorted.length
+    ? sorted.map(
+        (t, i) => `${medals[i] ?? `${i + 1}.`} **${t.teamName}** — ${formatGp(t.currentPot)} gp`,
+      )
+    : ['No teams recorded'];
+
+  const embed = {
+    title: `🏁  ${eventName}  🏁`,
+    color: 0x00b4d8,
+    description: [
+      `# 🏆 The Gielinor Rush has ended!`,
+      ``,
+      `> The dust has settled and the scores are in.`,
+      `> Thanks to everyone who competed, it was a sweet-ass race!`,
+      ``,
+      `## **[🔗 View Final Results →](${eventUrl})**`,
+    ].join('\n'),
+    fields: [
+      { name: '​', value: '────────────────────────────────────────' },
+      { name: '🏆  Final Standings', value: standingsLines.join('\n') },
+      { name: '​', value: '────────────────────────────────────────' },
+    ],
+    footer: { text: '🎉 Thanks for playing, see you next time!' },
+    timestamp: new Date().toISOString(),
+  };
+
+  const sent = await sendToChannels(channels, { embeds: [embed] });
+  return { success: sent > 0, channelsSent: sent };
+}
+
+async function sendStartMessage(guildId, eventId, eventName, teams) {
+  const channels = await getEventChannels(guildId, eventId);
+  if (channels.length === 0)
+    return { success: false, error: 'No channels found with event ID in topic' };
 
   const eventUrl = `${FRONTEND_URL}/gielinor-rush/${eventId}`;
   const teamList = teams.map((t) => `\`${t.teamName}\``).join('  ·  ') || 'TBD';
 
   const embed = {
-    title: `🏆  ${eventName}  🏆`,
+    title: `🚨  ${eventName}  🚨`,
     color: 0xf0c040,
     description: [
-      `# 🗺️ The Gielinor Rush is live!`,
+      `# 🗺️ The Gielinor Rush is NOW LIVE!`,
       ``,
       `> Grab your gear, rally your crew, and race across Gielinor.`,
-      `> The map is open and glory awaits. May the best team win!!`,
+      `> The map is open, submissions are open, and glory awaits.`,
+      `> May the best team win!!`,
       ``,
       `## **[🔗 Open Event Overview →](${eventUrl})**`,
     ].join('\n'),
     fields: [
-      { name: '🆔  Event ID', value: `\`\`\`${eventId}\`\`\``, inline: true },
       { name: '👥  Competing Teams', value: teamList, inline: true },
       { name: '​', value: '────────────────────────────────────────' },
       {
@@ -87,44 +205,4 @@ async function sendLaunchMessage(guildId, eventId, eventName, teams) {
   return { success: sent > 0, channelsSent: sent };
 }
 
-async function sendCompleteMessage(guildId, eventId, eventName, teams) {
-  const channels = await getEventChannels(guildId, eventId);
-  if (channels.length === 0) return { success: false, error: 'No channels found with event ID in topic' };
-
-  const eventUrl = `${FRONTEND_URL}/gielinor-rush/${eventId}`;
-  const medals = ['🥇', '🥈', '🥉'];
-  const formatGp = (gpStr) => BigInt(gpStr || '0').toLocaleString();
-
-  const sorted = [...(teams || [])].sort((a, b) =>
-    BigInt(b.currentPot || '0') > BigInt(a.currentPot || '0') ? 1 : -1
-  );
-
-  const standingsLines = sorted.length
-    ? sorted.map((t, i) => `${medals[i] ?? `${i + 1}.`} **${t.teamName}** — ${formatGp(t.currentPot)} gp`)
-    : ['No teams recorded'];
-
-  const embed = {
-    title: `🏁  ${eventName}  🏁`,
-    color: 0x00b4d8,
-    description: [
-      `# 🏆 The Gielinor Rush has ended!`,
-      ``,
-      `> The dust has settled and the scores are in.`,
-      `> Thanks to everyone who competed, it was a sweet-ass race!`,
-      ``,
-      `## **[🔗 View Final Results →](${eventUrl})**`,
-    ].join('\n'),
-    fields: [
-      { name: '​', value: '────────────────────────────────────────' },
-      { name: '🏆  Final Standings', value: standingsLines.join('\n') },
-      { name: '​', value: '────────────────────────────────────────' },
-    ],
-    footer: { text: '🎉 Thanks for playing, see you next time!' },
-    timestamp: new Date().toISOString(),
-  };
-
-  const sent = await sendToChannels(channels, { embeds: [embed] });
-  return { success: sent > 0, channelsSent: sent };
-}
-
-module.exports = { registerClient, sendLaunchMessage, sendCompleteMessage };
+module.exports = { registerClient, sendLaunchMessage, sendStartMessage, sendCompleteMessage };
